Building ECS framework
SubShip class partially filled out. ComponentTypes are being filled in. Frameworked out the entire project but realized to get this rolling it needs to be done incrementally. Removed all weapon-related things until ships are in and able to fly around.

SpawnSubShipSystem is loosely based on the ECS sample repo. Needs more work.