Refactor. Working through movement.
Decided any test methods will have a TEST_ suffix. Quaternions are very foreign. Played around with them and found quaternion multiplication is overloaded with Vector3 so Q * Vector3.forward yields a vector in the forward direction. Thrust can then be applied to it to change the magnitude.

Need to do some restructuring. SpawnSubShipSystem needs to be broken apart.