Orientation updates applied. Settings & Common created.
Entities should now apply orientation updates. When an entity applies a thrust with any orientation it should add that to the previous vector to obtain a new vector.

Settings created. Only one in there at the moment. Common created to find magnitude of floats, specifically for the newly created Settings.MaxShipSpeed.