Added ship reference
Needed to be able to assign the camera to a particular ship. This proved extremely difficult because Unity's Entity.Index is "useless". GetAllEntities()[goal_index] is undefined because internally unity shuffles Entities around. You have to walk the whole array to find the Entity where Index == goal_index. This was not a reasonable thing to do so instead it was determined that a reference to the Entity should be kept.

It works though! Camera is locked onto the Entity's position, it just took a long time to figure out.