Added look mechanic
Realized the camera game object was being found every OnUpdate call. Made a reference variable and set it on Start(). Added ability to look around in the ship. This is intended to be a VR game so look can be coupled with turning but needs to be kept as a separate thing.