Turning ship with the mouse
Working on getting the ship to turn with the mouse. It does but it's not quite right. I'm pushing because it does rotate the ship (even though it's not 100% right yet). Also, the ship moves in the direction the camera is facing which is great. It's come a long way :+1: