Camera "corrected"
Camera now correctly follows first person (cockpit) view. Took the easy road and projected the cockpit onto the orientation. The correct way is going to be with a vector rotation about the normal. Found a bug: can't do loops. The rotation breaks when doing a loop, probably due to the way yaw/pitch are setup. Need to find a better way of handling translation of mouse input to camera/ship movement.