Projectile collision test implementation complete
Tested an implementation that would run n^2 on projectiles to see if there were any collisions. "Pure" ECS does not have proper collision completed. Created a collision detector for two spheres.

Current plan is to have a two-part collision detection system. All ships will have a large "collision sphere". If a projectile or ship collides with that sphere we can then descend into the collision boxes of that ship to check for collision.

For this arm of the project (prep for Thesis) there will only be the sphere collider for the ships (no sub-boxes).