Projectiles damage and destroy ships!
Gave ships a collision sphere. If a projectile enters that sphere the ship is damaged. When ship health drops to zero the ship entitiy is destroyed. Projectile entities are destroyed on impact or when they travel a set distance from the origin. This keeps FPS in the 300 range. Performant enough.