Space partitioning in work
Since ECS doesn't have colliders it is necessary to efficiently know distances to objects. It was decided that space will be cut into octants. The entire space will be subdivided into 729 blocks (9x9x9), with each octant being 5x5x5 with an overlap on 3 planes. Each octant can be recursively built as a "new space" to reduce the number of entities in that space.

It is "expensive" to run distance calculations against each entities for all other entities. This segmentation of space allows for smaller groupings of entities so shortest distance is faster to calculate. This means the space will have to be built each frame but segmentation of space is order O(N), and distance calculations will be at most O(N^2) on each octant.

Octant recursive calls will be determined by an entity threshold count. If an octant contains more than X entities it will be subdivided into octants. Since subdivision is O(N) this isn't prohibitive.