Fixed resolution issue
Couldn't get it to play in 2K resolution. Found a solution. Will eventually need a menu but the foundation that was set today will be good even after the menu is made. Menu changes will change PlayerSettings, which will then update with SetResolution. SetResolution is likely to be moved down the road.